+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - gradle: support different java home in wrapper scripts
+- sos-report: add more controller json(resource-definitions, snapshots) and zfs list snapshots
 
 ## [1.30.2] - 2024-12-18